+++ e2e/playwright/regression/@existing-airgapped-install-minimal/test.spec.ts
@@ -26,6 +26,7 @@ import {
   cliAirgapInstall,
   installVeleroHostPath,
   validateDashboardGraphs,
+  validateGenerateSupportBundleUi,
   updateConfig,
   validateVersionMinimalRBACPreflights,
   validateCurrentVersionCard,
@@ -144,7 +145,8 @@ test('type=existing cluster, env=airgapped, phase=new install, rbac=minimal rbac
     registryInfo
   );
 
-  // TODO NOW: generate/validate support bundle
+  // Support bundle
+  await validateGenerateSupportBundleUi(page, expect, constants.IS_AIRGAPPED);
 
   // Config update and version history checks
   await updateConfig(page, expect);
@@ -162,7 +164,7 @@ test('type=existing cluster, env=airgapped, phase=new install, rbac=minimal rbac
   await validateVersionDiff(page, expect, 3, 2);
   await deployNewVersion(page, expect, 3, 'License Change', constants.IS_MINIMAL_RBAC);
 
-  // // Snapshot validation
+  // Snapshot validation
   await validateSnapshotsHostPathConfig(page, expect);
   await validateAutomaticFullSnapshots(page, expect);
   await validateAutomaticPartialSnapshots(page, expect);
