+++ e2e/playwright/shared-core/index.ts
@@ -1,0 +1,1 @@
+export * from './support-bundle';
